refactor: move classes/interfaces
