+++ src/ByteSync.Client/Business/Comparisons/DescriptionBuilderFactory.cs
@@ -2,6 +2,7 @@
 using ByteSync.Interfaces.Business.Actions;
 using ByteSync.Interfaces.Converters;
 using ByteSync.Interfaces.Factories;
+using ByteSync.Interfaces.Services.Localizations;
 using ByteSync.Services.Comparisons.DescriptionBuilders;
 
 namespace ByteSync.Business.Comparisons;
