+++ src/ByteSync.Client/Factories/ViewModels/DateAndInventoryPartsViewModelFactory.cs
@@ -1,5 +1,6 @@
 ﻿using ByteSync.Interfaces;
 using ByteSync.Interfaces.Factories.ViewModels;
+using ByteSync.Interfaces.Services.Localizations;
 using ByteSync.Interfaces.Services.Sessions;
 using ByteSync.Models.Inventories;
 using ByteSync.ViewModels.Sessions.Comparisons.Results;
