+++ src/ByteSync.Client/Interfaces/Services/Localizations/IAtomicActionValidationFailureReasonService.cs
@@ -1,0 +1,8 @@
+﻿using ByteSync.Business.Comparisons;
+
+namespace ByteSync.Interfaces.Services.Localizations;
+
+public interface IAtomicActionValidationFailureReasonService
+{
+    string GetLocalizedMessage(AtomicActionValidationFailureReason reason);
+}
