+++ src/ByteSync.Client/Services/Communications/WebAccessor.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using ByteSync.Interfaces;
 using ByteSync.Interfaces.Controls.Communications;
+using ByteSync.Interfaces.Services.Localizations;
 
 namespace ByteSync.Services.Communications;
