+++ src/ByteSync.Client/Services/Comparisons/DescriptionBuilders/AbstractDescriptionBuilder.cs
@@ -2,6 +2,7 @@
 using ByteSync.Assets.Resources;
 using ByteSync.Common.Business.Actions;
 using ByteSync.Interfaces;
+using ByteSync.Interfaces.Services.Localizations;
 
 namespace ByteSync.Services.Comparisons.DescriptionBuilders;
