+++ src/ByteSync.Client/Services/Comparisons/DescriptionBuilders/AtomicConditionDescriptionBuilder.cs
@@ -4,6 +4,7 @@ using ByteSync.Business.Comparisons;
 using ByteSync.Interfaces;
 using ByteSync.Interfaces.Business.Actions;
 using ByteSync.Interfaces.Converters;
+using ByteSync.Interfaces.Services.Localizations;
 
 namespace ByteSync.Services.Comparisons.DescriptionBuilders;
