+++ src/ByteSync.Client/Services/Comparisons/DescriptionBuilders/SharedAtomicActionDescriptionBuilder.cs
@@ -2,6 +2,7 @@
 using ByteSync.Business.Actions.Shared;
 using ByteSync.Common.Business.Actions;
 using ByteSync.Interfaces;
+using ByteSync.Interfaces.Services.Localizations;
 
 namespace ByteSync.Services.Comparisons.DescriptionBuilders;
