+++ src/ByteSync.Client/Services/Comparisons/DescriptionBuilders/SynchronizationRuleDescriptionBuilder.cs
@@ -4,6 +4,7 @@ using ByteSync.Business.Comparisons;
 using ByteSync.Interfaces;
 using ByteSync.Interfaces.Business.Actions;
 using ByteSync.Interfaces.Factories;
+using ByteSync.Interfaces.Services.Localizations;
 
 namespace ByteSync.Services.Comparisons.DescriptionBuilders;
