+++ src/ByteSync.Client/Services/Converters/BoolToDeploymentModeConverter.cs
@@ -4,6 +4,7 @@ using Avalonia.Controls;
 using Avalonia.Data.Converters;
 using ByteSync.Assets.Resources;
 using ByteSync.Interfaces;
+using ByteSync.Interfaces.Services.Localizations;
 
 namespace ByteSync.Services.Converters;
