+++ src/ByteSync.Client/Services/Converters/FormatKbSizeConverter.cs
@@ -6,6 +6,7 @@ using ByteSync.Assets.Resources;
 using ByteSync.Common.Helpers;
 using ByteSync.Interfaces;
 using ByteSync.Interfaces.Converters;
+using ByteSync.Interfaces.Services.Localizations;
 
 namespace ByteSync.Services.Converters;
