+++ src/ByteSync.Client/Services/Converters/SizeUnitConverter.cs
@@ -4,6 +4,7 @@ using ByteSync.Assets.Resources;
 using ByteSync.Common.Business.Misc;
 using ByteSync.Interfaces;
 using ByteSync.Interfaces.Converters;
+using ByteSync.Interfaces.Services.Localizations;
 
 namespace ByteSync.Services.Converters;
