+++ src/ByteSync.Client/ViewModels/Misc/ErrorViewModel.cs
@@ -4,6 +4,7 @@ using Avalonia.Controls.Mixins;
 using ByteSync.Assets.Resources;
 using ByteSync.Common.Helpers;
 using ByteSync.Interfaces;
+using ByteSync.Interfaces.Services.Localizations;
 using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
