+++ src/ByteSync.Client/ViewModels/Sessions/Comparisons/Actions/SynchronizationRuleGlobalViewModel.cs
@@ -10,6 +10,7 @@ using ByteSync.Interfaces;
 using ByteSync.Interfaces.Controls.Synchronizations;
 using ByteSync.Interfaces.Dialogs;
 using ByteSync.Interfaces.Factories.ViewModels;
+using ByteSync.Interfaces.Services.Localizations;
 using ByteSync.Interfaces.Services.Sessions;
 using ByteSync.ViewModels.Misc;
 using ByteSync.ViewModels.Sessions.Comparisons.Actions.Misc;
