+++ src/ByteSync.Client/ViewModels/Sessions/DataNodes/DataNodeHeaderViewModel.cs
@@ -12,6 +12,7 @@ using ByteSync.Interfaces;
 using ByteSync.Interfaces.Controls.Inventories;
 using ByteSync.Interfaces.Controls.Themes;
 using ByteSync.Interfaces.Repositories;
+using ByteSync.Interfaces.Services.Localizations;
 using ByteSync.Interfaces.Services.Sessions;
 using ByteSync.ViewModels.Misc;
 using DynamicData;
