+++ src/ByteSync.Client/ViewModels/Sessions/Managing/AnalysisModeViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using ByteSync.Business.Sessions;
 using ByteSync.Interfaces;
+using ByteSync.Interfaces.Services.Localizations;
 using ReactiveUI.Fody.Helpers;
 
 namespace ByteSync.ViewModels.Sessions.Managing;
