+++ tests/ByteSync.Client.Tests/Services/Converters/ItemCountToLocalizedTextConverterTests.cs
@@ -1,5 +1,6 @@
 using System.Globalization;
 using ByteSync.Interfaces;
+using ByteSync.Interfaces.Services.Localizations;
 using ByteSync.Services.Converters;
 using FluentAssertions;
 using Moq;
