+++ tests/ByteSync.Client.Tests/ViewModels/Sessions/Comparisons/Actions/TargetedActionGlobalViewModelTests.cs
@@ -9,6 +9,7 @@ using ByteSync.Interfaces;
 using ByteSync.Interfaces.Controls.Comparisons;
 using ByteSync.Interfaces.Dialogs;
 using ByteSync.Interfaces.Factories.ViewModels;
+using ByteSync.Interfaces.Services.Localizations;
 using ByteSync.Models.Comparisons.Result;
 using ByteSync.Services.Localizations;
 using ByteSync.TestsCommon;
